task 0: prints all elements of nodes

+++ 0x13-more_singly_linked_lists/lists.h
@@ -12,7 +12,7 @@
  */
 typedef struct list
 {
-	int n;
+	unsigned long int n;
 	struct list *next;
 } listint_t;
